requirements.txt
infi.systray
pyserial
